move to 2.0 auth login-form component
