+++ ui/app/components/auth/form/okta.hbs
@@ -1,13 +1,21 @@
-<form {{on "submit" this.login}}>
-  {{yield}}
+{{#if this.pollForOktaNumberChallenge.isRunning}}
+  <OktaNumberChallenge
+    @correctAnswer={{this.challengeAnswer}}
+    @hasError={{this.authError}}
+    @onReturnToLogin={{this.onCancel}}
+  />
+{{else}}
+  <form {{on "submit" this.login}}>
+    {{yield}}
 
-  <div class="margin-24">
-    {{#each this.showFields as |field|}}
-      <Hds::Form::TextInput::Field @type={{this.maybeMask field}} name={{field}} class="has-bottom-margin-m" as |F|>
-        <F.Label>{{capitalize field}}</F.Label>
-      </Hds::Form::TextInput::Field>
-    {{/each}}
+    <div class="margin-24">
+      {{#each this.showFields as |field|}}
+        <Hds::Form::TextInput::Field @type={{this.maybeMask field}} name={{field}} class="has-bottom-margin-m" as |F|>
+          <F.Label>{{capitalize field}}</F.Label>
+        </Hds::Form::TextInput::Field>
+      {{/each}}
 
-    <Hds::Button @text="Sign in" @isFullWidth={{true}} type="submit" class="has-top-margin-m has-bottom-margin-m" />
-  </div>
-</form>
+      <Hds::Button @text="Sign in" @isFullWidth={{true}} type="submit" class="has-top-margin-m has-bottom-margin-m" />
+    </div>
+  </form>
+{{/if}}
