+++ ui/app/components/auth/login-form-old.hbs
@@ -1,0 +1,23 @@
+{{!
+  Copyright (c) HashiCorp, Inc.
+  SPDX-License-Identifier: BUSL-1.1
+~}}
+
+{{#if this.waitingForOktaNumberChallenge}}
+  <OktaNumberChallenge
+    @correctAnswer={{this.oktaNumberChallengeAnswer}}
+    @hasError={{this.authError}}
+    @onReturnToLogin={{this.onCancel}}
+  />
+{{else}}
+  <AuthForm
+    @wrappedToken={{@wrappedToken}}
+    @cluster={{@cluster}}
+    @namespace={{@namespace}}
+    @selectedAuth={{@selectedAuth}}
+    @error={{this.authError}}
+    @performAuth={{this.performAuth}}
+    @authIsRunning={{this.authenticate.isRunning}}
+    @delayIsIdle={{this.delayAuthMessageReminder.isIdle}}
+  />
+{{/if}}
