+++ ui/app/components/auth/page.hbs
@@ -55,81 +55,14 @@
           @onError={{fn (mut this.mfaErrors)}}
         />
       {{else}}
-        {{! This would be a separate component, leaving here just to show auth component POC }}
-        {{#if (or this.selectedAuthMethod this.authTabs)}}
-          <Auth::FormTemplate
-            @wrappedToken={{@wrappedToken}}
-            @cluster={{@cluster}}
-            {{! set namespace in service instead of pass to child? }}
-            {{!-- @namespace={{@namespaceQueryParam}} --}}
-            @authType={{this.selectedAuthMethod}}
-            @onSuccess={{this.onAuthResponse}}
-          >
-            <:namespace>
-              {{#if this.version.isEnterprise}}
-                <Auth::NamespaceInput
-                  @updateNamespace={{this.handleNamespaceUpdate}}
-                  @value={{@namespace}}
-                  @hvdManagedNamespace={{this.flags.hvdManagedNamespaceRoot}}
-                  @disabled={{@oidcProviderQueryParam}}
-                />
-              {{/if}}
-            </:namespace>
-            <:options>
-              {{! "Back" button shows when user has clicked "Sign in with other methods" }}
-              {{#if (and this.showAllMethods this.authTabs)}}
-                <Hds::Button
-                  @text="Back"
-                  {{on "click" (fn (mut this.showAllMethods) false)}}
-                  class="has-left-margin-l"
-                  @color="tertiary"
-                  @icon="arrow-left"
-                />
-              {{/if}}
-            </:options>
-            <:tabs>
-              {{#if (and this.authTabs (not this.showAllMethods))}}
-                <Auth::Tabs
-                  @authTabs={{this.authTabs}}
-                  @displayNameHelper={{this.displayName}}
-                  @onTabClick={{this.handleAuthSelect}}
-                  @selectedAuthMethod={{this.selectedAuthMethod}}
-                  @selectedTabIndex={{this.selectedTabIndex}}
-                />
-              {{else}}
-                <div class="has-left-margin-l has-bottom-margin-l has-right-margin-l">
-                  {{! select changes this.selectedAuthType }}
-                  <Hds::Form::Select::Field name="selectedAuthMethod" {{on "input" this.handleAuthSelect}} as |F|>
-                    <F.Label class="has-top-padding-l">Auth method</F.Label>
-                    <F.Options>
-                      {{#each this.availableMethodTypes as |type|}}
-                        <option selected={{eq this.selectedAuthMethod type}} value={{type}}>
-                          {{this.displayName type}}
-                        </option>
-                      {{/each}}
-                    </F.Options>
-                  </Hds::Form::Select::Field>
-                  {{#if (not-eq @selectedAuthMethod "token")}}
-                    <Hds::Form::TextInput::Field name="path" as |F|>
-                      <F.Label class="has-top-padding-l">Mount path</F.Label>
-                    </Hds::Form::TextInput::Field>
-                  {{/if}}
-                </div>
-              {{/if}}
-            </:tabs>
-          </Auth::FormTemplate>
-        {{/if}}
-        {{#unless this.showAllMethods}}
-          <Hds::Button
-            @isInline={{true}}
-            @text="Sign in with other methods"
-            @color="tertiary"
-            @icon="arrow-right"
-            @iconPosition="trailing"
-            {{on "click" (fn (mut this.showAllMethods) true)}}
-            class="has-left-margin-m has-bottom-margin-l"
-          />
-        {{/unless}}
+        <Auth::LoginForm
+          @wrappedToken={{@wrappedToken}}
+          @cluster={{@cluster}}
+          @handleNamespaceUpdate={{this.handleNamespaceUpdate}}
+          {{! idea: set namespace in service instead of pass to child and manage state manually}}
+          @namespace={{@namespaceQueryParam}}
+          @onSuccess={{this.onAuthResponse}}
+        />
       {{/if}}
     </:content>
